% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{absences}
\alias{absences}
\title{Virtual species absences}
\format{A data frame with 250 rows and 3 variables:
\describe{
  \item{sp}{virtual species names}
  \item{x}{longitude of species occurrences}
  \item{y}{latitude of species occurrences}
  ...
}}
\usage{
absences
}
\description{
A dataset containing absences of 5 virtual species
}
\keyword{datasets}
