% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSDM_Posteriori.R
\name{MSDM_Posteriori}
\alias{MSDM_Posteriori}
\title{Methods to correct overprediction of species distribution models based on occurrences and suitability patterns of species.}
\usage{
MSDM_Posteriori(records, absences, x = NA, y = NA, sp = NA,
  method = c("OBR", "PRES", "LQ", "MCP", "BMCP"), dirraster = NULL,
  threshold = c("kappa", "spec_sens", "no_omission", "prevalence",
  "equal_sens_spec", "sensitivty"), buffer = c("single", "species_specific"),
  dirsave = NULL)
}
\arguments{
\item{records}{data.frame. A database with geographical coordinates of species presences used to create species distribution models.}

\item{absences}{data.frame. A database with geographical coordinates of species absences used to create species distribution models.}

\item{x}{character. Column name with longitude values. This name must be the same for presences and absences database.}

\item{y}{character. Column name with latitude values. This name must be the same for presences and absences database.}

\item{sp}{character. Column name with species names. Species names must be the same for presences, absences, and raster layer (i.e. species distribution) databases.}

\item{method}{character. A character string indicating which MSDM method must be used create.}

\item{dirraster}{character. A character string indicating the directory where are species raster files, i.e. species distribution models.
Raster layer must be in geotiff format}

\item{threshold}{character. Select type of threshold (kappa, spec_sens, no_omission, prevalence, equal_sens_spec, sensitivty)
to get binary models (see \code{\link[dismo]{threshold}} help of dismo package for further information about differt thresholds). Default threshold value is "equal_sens_spec", it is the threshold at which sensitivity and specificity are equal.}

\item{buffer}{character. Type o buffer width use in BMCP approach.}

\item{dirsave}{character. A character string indicating the directory where result must be saved.}
}
\value{
This function save raster files (with geotiff format) with continuous and binary species raster files separated in CONr and BINr folders respectively.
}
\description{
Methods to correct overprediction of species distribution models based on occurrences and suitability patterns of species.
}
\details{
Abbreviation list

\itemize{
\item SDM: species distribution model
\item l: suitability patches that intercept species occurrences
\item k: suitability patches that do not intercept species occurrences
\item T: threshold distances used to select suitability patches
}


All the methods used by this function use raster layer with
the suitability values of a species distribution models (SDMs). After applied a threshold to binarize SDMs
 (see 'threshold' arguments), function create a binarized raster layer with suitable and unsuitable patches.


OBR(Occurrences based restriction)-
The method assumes that suitable patches intercepting species occurrences (l)
are likely a part of species distributions than suitable patches that do not
intercept any occurrence (k). Distance from all patches k species occurrences to the closest l
patch is calculated, later it is removed k patches that trespass a species-specific
distance threshold from SDMs models. This threshold (T) is calculated as the
maximum distance in a vector of minimal pairwise distances between occurrences.
Whenever a suitable pixel is within a k patch distant from the closest l in less than T,
the suitability of the pixel was reduced to zero. We assumed that this simple threshold
is a surrogate of the species-specific dispersal ability. If T is low, either the species
has been sampled throughout its distribution, or the species is geographically restricted,
justifying a narrow inclusion of k patches.

PRES (Only occurrences based restriction). This is a more restrictive variant of the OBR method. It only retains those pixels in suitability patches intercepting occurrences (k)

LQ (Lower Quantile). This method is similar to the OBR method, except by the
procedure to define a distance threshold to withdrawn k patches, which is the
lower quartile distance between k patches to the closest l patch. Whenever a suitable
pixel is within a k patch, i.e. not within this lower quartile, the suitability of the
pixel is reduced to zero. This means that 75% of k patches were withdrawn from the model.

MCP (Minimum Convex Polygon). Compiled and adapted from
Kremen et al. (2008), this method excludes from SDMs climate suitable
pixels that do not intercept a minimum convex polygon,
with interior angles smaller than 180, enclosing all occurrences of a species.

BMCP (Buffered Minimum Convex Polygon). Compiled and adapted
from Kremen et al. (2008), it is alike the MCP except by the inclusion of a
buffer zone surrounding minimum convex polygons. When used with the "single" options for buffer algument
function will ask for a value in km to be used as the buffer with. When used "species_specific" a buffer will be calculated for each species based on the presences occurrences patterns, assuming as buffer width
the maximum distance in a vector of minimal pairwise distances between occurrences.
}
\examples{

require(MSDM)
require(raster)
data("occurrences")
data("absences")

dir_raster <- system.file("extdata", package="MSDM")
# List of raster layer with species suitability
list.files(dir_raster, full.names = TRUE)

# Create a temporary MSDM folder
tmdir <- tempdir()
tmdir
dir.create(file.path(tmdir,"MSDM"))
tmdir <- file.path( tmdir,"MSDM")
tmdir

# MCP method----
MSDM_Posteriori(records=occurrences, absences=absences,
                x="x", y="y", sp="sp", method="MCP",
                dirraster = dir_raster, threshold = "spec_sens",
                dirsave = tmdir)

d <- list.dirs(tmdir, recursive = FALSE)
# Categorical models corrected by MCP methods
cat_mcp <- stack(list.files(d[1], full.names = TRUE))
plot(cat_mcp)
# Continuous models corrected by MCP methods
con_mcp <- stack(list.files(d[2], full.names = TRUE))
plot(con_mcp)


# BMCP method with a single buffer for all species----
MSDM_Posteriori(records=occurrences, absences=absences,
                x="x", y="y", sp="sp", method="BMCP", buffer="single",
                dirsave = tmdir)

d <- list.dirs(tmdir, recursive = FALSE)
# Categorical models corrected by MCP methods
cat_bmcp <- stack(list.files(d[1], full.names = TRUE))
plot(cat_bmcp)
# Continuous models corrected by MCP methods
con_bmcp <- stack(list.files(d[2], full.names = TRUE))
plot(con_bmcp)

# LQ method----
MSDM_Posteriori(records=occurrences, absences=absences,
                x="x", y="y", sp="sp", method="LQ",
                dirraster = dir_raster, threshold = "spec_sens",
                dirsave = tmdir)

d <- list.dirs(tmdir, recursive = FALSE)

# Categorical models corrected by LQ methods
cat_lq <- stack(list.files(d[1], full.names = TRUE))
plot(cat_lq)
# Continuous models corrected by LQ methods
con_lq <- stack(list.files(d[2], full.names = TRUE))
plot(con_lq)


# OBR method----
MSDM_Posteriori(records=occurrences, absences=absences,
                x="x", y="y", sp="sp", method="OBR",
                dirraster = dir_raster, threshold = "spec_sens",
                dirsave = tmdir)

d <- list.dirs(tmdir, recursive = FALSE)

# Categorical models corrected by OBR methods
cat_obr <- stack(list.files(d[1], full.names = TRUE))
plot(cat_obr)
# Continuous models corrected by OBR methods
con_obr <- stack(list.files(d[2], full.names = TRUE))
plot(con_obr)


# PRES method----
MSDM_Posteriori(records=occurrences, absences=absences,
                x="x", y="y", sp="sp", method="PRES",
                dirraster = dir_raster, threshold = "spec_sens",
                dirsave = tmdir)

d <- list.dirs(tmdir, recursive = FALSE)

# Categorical models corrected by PRES methods
cat_pres <- stack(list.files(d[1], full.names = TRUE))
plot(cat_pres)
# Continuous models corrected by PRES methods
con_pres <- stack(list.files(d[2], full.names = TRUE))
plot(con_pres)


}
\references{
\itemize{
\item Kremen, C., Cameron, A., Moilanen, A., Phillips, S. J., Thomas, C. D.,
Beentje, H., . Zjhra, M. L. (2008). Aligning Conservation Priorities Across
Taxa in Madagascar with High-Resolution Planning Tools. Science, 320(5873),
222-226. doi:10.1126/science.1155193
}
}
\seealso{
\code{\link{MSDM_Priori}}
}
