% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSDM_Priori.R
\name{MSDM_Priori}
\alias{MSDM_Priori}
\title{Create spatial predictor variables to correct overprediction of species distribution models}
\usage{
MSDM_Priori(records, x = NA, y = NA, sp = NA, method = c("XY", "MIN",
  "CML", "KER"), rasterlayer = NULL, dirsave = NULL)
}
\arguments{
\item{records}{data.frame. A database with geographical coordinates of species presences used to create species distribution models.}

\item{x}{character. Column name with longitude values.}

\item{y}{character. Column name with latitude values.}

\item{sp}{character. Column name with species names.}

\item{method}{character. A character string indicating which MSDM method must be used create.}

\item{rasterlayer}{raster object. A raster, stack or brick object that will be used to construct species distribution models. This object will be used as a basis for creating MSDM variables with the same resolution, extent and pattern of empty cell that the environmental variables.}

\item{dirsave}{character. A character string indicating the directory where result must be saved.}
}
\value{
This function save raster files (with geotiff format) in a folder named with the MSDM method. Such raster have to be used as covariate together with environmental variables at moment to construct species distribution models. A pair of raster layer are created for all species set for XY method, on the contrary, CML and KER create a specific raster layer for each species.
}
\description{
Create spatial predictor variables to correct overprediction of species distribution models
}
\details{
XY (Latlong method). It assumes that spatial structure can partially explain
species distribution (Bahn & Mcgill, 2007). Two raster layers will be created,
containing the latitude and longitude of pixels in decimal degrees respectively.
These raster layers should be included as covariates with the environmental layers to construct species distribution models.


CML (Cumulative distance method). Compiled and adapted from Allouche et al., (2008), it assumes that
pixels closer to presences are likely included in species distributions.
A raster layer will be created containing the sum of Euclidian geographic distances
from each pixel to all occurrences of a species. Obtained values are normalized to
vary from zero to one. This raster layer should be included as covariates with the environmental layers to construct species distribution models.


KER (Kernel method). Also compiled and adapted from Allouche et al. (2008), this method,
alike CML, assumes that pixels located in areas with higher density of occurrences are
likely included in the actual species distribution. A raster layer was created containing
the Gaussian value based in the density of occurrences of a species. Standard deviation
of the Gaussian distribution was the maximum value in a vector of minimum distances between
pairs of occurrences of a species. Gaussian values are normalized to vary from zero to one.
This raster layer should be included as covariates with the environmental layers to construct species distribution models.
}
\examples{
library(MSDM)
library(raster)

data("rlayer")
data("occurrences")

plot(rlayer)
head(occurrences)

tmdir <- tempdir()
tmdir # temporal directory where will be saves raster layers

# XY method----
MSDM_Priori(records = occurrences,
            x = "x", y = "y", sp = "sp", method = "XY",
            rasterlayer = rlayer, dirsave = tmdir)

# open directory were raster were saved
rdir <- paste(tmdir, "MSDM_XY", sep = '/')
rdir
shell.exec(rdir)

# read new MSDM layers
new_var <- list.files(rdir, pattern = ".tif", full.names = TRUE)
new_var <- stack(new_var)
plot(new_var)

# CML method----
MSDM_Priori(records = occurrences,
            x = "x", y = "y", sp = "sp", method = "CML",
            rasterlayer = rlayer, dirsave = tmdir)

# open directory were raster were saved
rdir <- paste(tmdir, "MSDM_CML", sep = '/')
rdir
shell.exec(rdir)

# read new MSDM layers
new_var <- list.files(rdir, pattern = ".tif", full.names = TRUE)
new_var <- stack(new_var)
plot(new_var)

# MIN method----
MSDM_Priori(records = occurrences,
            x = "x", y = "y", sp = "sp", method = "MIN",
            rasterlayer = rlayer, dirsave = tmdir)

# open directory were raster were saved
rdir <- paste(tmdir, "MSDM_MIN", sep = '/')
rdir
shell.exec(rdir)

# read new MSDM layers
new_var <- list.files(rdir, pattern = ".tif", full.names = TRUE)
new_var <- stack(new_var)
plot(new_var)


# KER methods----
MSDM_Priori(records = occurrences,
            x = "x", y = "y", sp = "sp", method = "KER",
            rasterlayer = rlayer, dirsave = tmdir)

# open directory were raster were saved
rdir <- paste(tmdir, "MSDM_KER", sep = '/')
rdir
shell.exec(rdir)

new_var <- list.files(rdir, pattern = ".tif", full.names = TRUE)
new_var <- stack(new_var)
plot(new_var)

}
\references{
\itemize{
\item Allouche, O., Steinitz, O., Rotem, D., Rosenfeld, A., & Kadmon, R. (2008a). Incorporating distance constraints into species distribution models. Journal of Applied Ecology, 45(2), 599-609. doi:10.1111/j.1365-2664.2007.01445.x
\item Bahn, V., & Mcgill, B. J. (2007). Can niche-based distribution models outperform spatial interpolation? Global Ecology and Biogeography, 16(6), 733-742. doi:10.1111/j.1466-8238.2007.00331.x
}
}
\seealso{
\code{\link{MSDM_Posteriori}}
}
