% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{occurrences}
\alias{occurrences}
\title{Virtual species occurrences.}
\format{
A data frame with 250 rows and 3 variables:
\describe{
  \item{sp}{virtual species names}
  \item{x}{longitude of species occurrences}
  \item{y}{latitude of species occurrences}
  ...
}
}
\usage{
occurrences
}
\description{
A dataset containing occurrences of five virtual species
}
\keyword{datasets}
