% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rlayer}
\alias{rlayer}
\title{A raster layer used to as expample in MSDM_Priori methods}
\format{A raster of Americas continent from southern USA to southern South America:}
\usage{
rlayer
}
\description{
A dataset containing absences of 5 virtual species
}
\examples{

require(raster)
data("rlayer")
plot(rlayer)
}
\keyword{datasets}
