% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sp_sdm}
\alias{sp_sdm}
\title{A dataset containing absences of five virtual species}
\format{
A raster of Americas from the southern United States to southern South America:
}
\usage{
sp_sdm
}
\description{
A dataset containing absences of five virtual species
}
\examples{

require(raster)
data("sp_sdm")
plot(sp_sdm)
}
\keyword{datasets}
