% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sp_sdm}
\alias{sp_sdm}
\title{A raster layer used as example in MSDM_Priori methods}
\format{
A raster of Americas continent from southern USA to southern South America:
}
\usage{
sp_sdm
}
\description{
A dataset containing absences of 5 virtual species
}
\examples{

require(raster)
data("sp_sdm")
plot(sp_sdm)
}
\keyword{datasets}
